[WV-2287] Moved ManageMyCandidates navigation to ManageMyCandidatesLanding.jsx and UI components to other files [TEAM REVIEW]
