+++ src/js/pages/ManageMyCandidates/SupporterTracking.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import styled from 'styled-components';
+import DesignTokenColors from '../../common/components/Style/DesignTokenColors';
+
+export default function SupporterTracking () {
+  return (
+    <Placeholder>
+      Tracking (coming soon)
+    </Placeholder>
+  );
+}
+
+const Placeholder = styled.div`
+  background: ${DesignTokenColors.neutralUI50};
+  border: 1px dashed ${DesignTokenColors.neutralUI300};
+  border-radius: 12px;
+  color: ${DesignTokenColors.neutralUI600};
+  padding: 24px;
+`;
